Refactoring Headers Middleware
